Remove: domain 디렉토리로 이동
