+++ src/main/java/org/example/fourchak/domain/coupon/repository/CouponRepository.java
@@ -2,7 +2,7 @@ package org.example.fourchak.domain.coupon.repository;
 
 import java.util.List;
 import org.example.fourchak.domain.coupon.entity.Coupon;
-import org.example.fourchak.store.entity.Store;
+import org.example.fourchak.domain.store.entity.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
