+++ src/main/java/org/example/fourchak/domain/reservation/controller/ReservationController.java
@@ -1,0 +1,11 @@
+package org.example.fourchak.domain.reservation.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class ReservationController {
+
+
+}
