+++ src/main/java/org/example/fourchak/domain/reservation/dto/requset/ReservationRequestDto.java
@@ -1,0 +1,12 @@
+package org.example.fourchak.domain.reservation.dto.requset;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+public class ReservationRequestDto {
+
+}
