+++ src/main/java/org/example/fourchak/domain/reservation/dto/response/ReservationResponseDto.java
@@ -1,0 +1,10 @@
+package org.example.fourchak.domain.reservation.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ReservationResponseDto {
+
+}
