+++ src/main/java/org/example/fourchak/domain/reservation/repository/ReservationRepository.java
@@ -1,0 +1,11 @@
+package org.example.fourchak.domain.reservation.repository;
+
+import org.example.fourchak.domain.reservation.entity.Reservation;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ReservationRepository extends JpaRepository<Reservation, Long> {
+
+
+}
