+++ src/main/java/org/example/fourchak/domain/reservation/service/ReservationServiceImpl.java
@@ -1,0 +1,8 @@
+package org.example.fourchak.domain.reservation.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class ReservationServiceImpl implements ReservationService {
+
+}
