+++ src/main/java/org/example/fourchak/domain/user/dto/request/UserRequest.java
@@ -1,0 +1,5 @@
+package org.example.fourchak.domain.user.dto.request;
+
+public class UserRequest {
+
+}
