+++ src/main/java/org/example/fourchak/domain/user/enums/UserRole.java
@@ -1,0 +1,6 @@
+package org.example.fourchak.domain.user.enums;
+
+public enum UserRole {
+    USER, ADMIN;
+
+}
