+++ src/main/java/org/example/fourchak/domain/user/service/UserService.java
@@ -1,0 +1,5 @@
+package org.example.fourchak.domain.user.service;
+
+public class UserService {
+
+}
