--- src/main/java/org/example/fourchak/reservation/controller/ReservationController.java
@@ -1,11 +1,0 @@
-package org.example.fourchak.reservation.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class ReservationController {
-
-
-}
