--- src/main/java/org/example/fourchak/reservation/dto/requset/ReservationRequestDto.java
@@ -1,12 +1,0 @@
-package org.example.fourchak.reservation.dto.requset;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@NoArgsConstructor
-public class ReservationRequestDto {
-
-}
