--- src/main/java/org/example/fourchak/reservation/dto/response/ReservationResponseDto.java
@@ -1,10 +1,0 @@
-package org.example.fourchak.reservation.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ReservationResponseDto {
-
-}
