--- src/main/java/org/example/fourchak/reservation/entity/Reservation.java
@@ -1,43 +1,0 @@
-package org.example.fourchak.reservation.entity;
-
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.example.fourchak.common.BaseEntity;
-
-@Getter
-@Entity
-@Table(name = "reservation")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Reservation extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private int peopleNumber;
-
-    @Column(nullable = false)
-    private LocalDateTime reservationTime;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "store_id")
-    private Store store;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-}
