--- src/main/java/org/example/fourchak/reservation/repository/ReservationRepository.java
@@ -1,11 +1,0 @@
-package org.example.fourchak.reservation.repository;
-
-import org.example.fourchak.reservation.entity.Reservation;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ReservationRepository extends JpaRepository<Reservation, Long> {
-
-
-}
