--- src/main/java/org/example/fourchak/reservation/service/ReservationService.java
@@ -1,5 +1,0 @@
-package org.example.fourchak.reservation.service;
-
-public interface ReservationService {
-
-}
