--- src/main/java/org/example/fourchak/reservation/service/ReservationServiceImpl.java
@@ -1,8 +1,0 @@
-package org.example.fourchak.reservation.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class ReservationServiceImpl implements ReservationService {
-
-}
