--- src/main/java/org/example/fourchak/user/controller/UserController.java
@@ -1,5 +1,0 @@
-package org.example.fourchak.user.controller;
-
-public class UserController {
-
-}
