--- src/main/java/org/example/fourchak/user/dto/request/UserRequest.java
@@ -1,5 +1,0 @@
-package org.example.fourchak.user.dto.request;
-
-public class UserRequest {
-
-}
