--- src/main/java/org/example/fourchak/user/dto/response/UserResponse.java
@@ -1,5 +1,0 @@
-package org.example.fourchak.user.dto.response;
-
-public class UserResponse {
-
-}
