--- src/main/java/org/example/fourchak/user/service/UserService.java
@@ -1,5 +1,0 @@
-package org.example.fourchak.user.service;
-
-public class UserService {
-
-}
